% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\alias{message_condition}
\alias{warning_condition}
\alias{error_condition}
\title{Create a condition object}
\usage{
condition(message, class = NULL, capture_call = TRUE, ...)

message_condition(message, class = NULL, capture_call = TRUE, ...)

warning_condition(message, class = NULL, capture_call = TRUE, ...)

error_condition(message, class = NULL, capture_call = TRUE, ...)
}
\arguments{
\item{message}{A message to display when the condition is raised}

\item{class}{Character vector of classes to append to the condition class}

\item{capture_call}{Logical. When raised, should the condition include the call?}

\item{...}{Additional arguments to append to the condition object}
}
\description{
Condition objects are classed objects which can be raised when desired.
\code{condition} objects are the most basic form; message, warning, and error
conditions are more specific. Additional classes can be added to make
condition handling easier and more targeted.
}
