% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/except.R
\name{except}
\alias{except}
\title{Catch conditions}
\usage{
except(expression, ...)
}
\arguments{
\item{expression}{Expression to evaluate within which to catch conditions}

\item{...}{Named arguments where names are condition classes to catch, and
arguments are functions to evaluate when that condition is raised.
Functions should take a single parameter, which will be passed the
condition object raised.}
}
\description{
Catch conditions
}
