% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise.R
\name{raise}
\alias{raise}
\title{Raise a condition}
\usage{
raise(condition)
}
\arguments{
\item{condition}{A \code{\link[=condition]{condition()}} to raise}
}
\description{
A generic function for raising any condition, including messages, warnings,
and errors. Conditions of other classes should define a method for this
generic.
}
